Renamed rest middlewares to use the short-url concept instead of the short-code concept
